Let sfdisk generate new disk id when initializing
This removes the old disk id from the info passed to sfdisk, forcing it
to generate a new disk id instead of preserving the old disk id when
updating the partition table. This allows removing the code that
randomizes the disk id after sfdisk updated the partition table. Also
note that this code was previously already redudundant because after
initializing there is more code that changes the disk id if it is
identical (to support the non-initializing flow, but that is also run
when initializing).