Support GPT partition tables
The sfdisk command already handles most of what is needed. This just
removes the last-lba entry from sfdisk partition info (which is based on
the source disk size and would otherwise limit the available destination
space, if it is omitted the entire destination disk is usable), and
makes change_disk_id support generating a longer UUID disk id when a GPT
table is in use.